+++ src/yt_game_records.c
@@ -63,7 +63,7 @@ yt_player_decode(struct yt_player *player, const struct yt_record *record)
 	player->lottery_plays = yt_record_get_number(record, YT_F105);
 	player->score = yt_record_get_number(record, YT_F109);
 	player->plasma = yt_record_get_number(record, YT_F113);
-	player->ports_owned = yt_record_get_number(record, YT_F117);
+	player->ports_owned = (int)yt_record_get_number(record, YT_F117);
 	player->ground_forces = yt_record_get_number(record, YT_F121);
 	player->cloak = yt_record_get_number(record, YT_F125);
 	player->mines = yt_record_get_number(record, YT_F129);
@@ -104,7 +104,7 @@ yt_player_encode(struct yt_player *player)
 	yt_record_set_number_if_changed(&player->record, YT_F109, player->score);
 	yt_record_set_number_if_changed(&player->record, YT_F113, player->plasma);
 	yt_record_set_number_if_changed(&player->record, YT_F117,
-	    player->ports_owned);
+	    (float)player->ports_owned);
 	yt_record_set_number_if_changed(&player->record, YT_F121,
 	    player->ground_forces);
 	yt_record_set_number_if_changed(&player->record, YT_F125, player->cloak);
